// src/shared/components/configurable-parameters/configurable-parameters.interface.ts
interface ConfigParametersId {
    id: string;
    autoHPOState: ConfigurableParametersGroupBasicParamsDTO['auto_hpo_state'];
}

type HPOOmitTypes = 'auto_hpo_state' | 'auto_hpo_value';

export interface SelectableGroupParams<T extends string | number, K extends ConfigGroupParametersDataTypesDTO>
    extends ConfigParametersId,
        Omit<SelectableGroupParamsDTO<T, K>, 'template_type' | 'data_type' | 'default_value' | HPOOmitTypes> {
    templateType: SelectableGroupParamsDTO<T, K>['template_type'];
    dataType: SelectableGroupParamsDTO<T, K>['data_type'];
    defaultValue: SelectableGroupParamsDTO<T, K>['default_value'];
    autoHPOValue: SelectableGroupParamsDTO<T, K>['auto_hpo_value'];
}

export interface NumberGroupParams
    extends ConfigParametersId,
        Omit<
            NumberGroupParamsDTO,
            'template_type' | 'data_type' | 'min_value' | 'max_value' | 'default_value' | HPOOmitTypes
        > {
    templateType: NumberGroupParamsDTO['template_type'];
    dataType: NumberGroupParamsDTO['data_type'];
    minValue: NumberGroupParamsDTO['min_value'];
    maxValue: NumberGroupParamsDTO['max_value'];
    defaultValue: NumberGroupParamsDTO['default_value'];
    autoHPOValue: NumberGroupParamsDTO['auto_hpo_value'];
}

export interface BooleanGroupParams
    extends ConfigParametersId,
        Omit<BooleanGroupParamsDTO, 'template_type' | 'data_type' | 'default_value' | HPOOmitTypes> {
    templateType: BooleanGroupParamsDTO['template_type'];
    dataType: BooleanGroupParamsDTO['data_type'];
    defaultValue: BooleanGroupParamsDTO['default_value'];
    autoHPOValue: BooleanGroupParamsDTO['auto_hpo_value'];
}

export type ConfigurableParametersParams =
    | SelectableGroupParams<string, 'string'>
    | SelectableGroupParams<number, 'integer' | 'float'>
    | BooleanGroupParams
    | NumberGroupParams;

export interface ConfigurableParametersGroups extends Omit<ConfigurableParametersGroupsDTO, 'parameters'> {
    id: string;
    parameters: ConfigurableParametersParams[];
}

interface EntityIdentifierGenericModel {
    modelStorageId: EntityModelIdentifierDTO['model_storage_id'];
    workspaceId: EntityModelIdentifierDTO['workspace_id'];
}

export interface EntityIdentifierHyperGroup
    extends Pick<EntityIdentifierHyperGroupDTO, 'type'>,
        EntityIdentifierGenericModel {
    groupName: EntityIdentifierHyperGroupDTO['group_name'];
}

type EntityIdentifierHyper = Pick<EntityIdentifierHyperDTO, 'type'> & EntityIdentifierGenericModel;

export type EntityModelIdentifier = EntityIdentifierHyperGroup | EntityIdentifierHyper;

export interface EntityTaskIdentifier extends Pick<EntityTaskIdentifierDTO, 'type' | 'component'> {
    workspaceId: EntityTaskIdentifierDTO['workspace_id'];
    taskId: EntityTaskIdentifierDTO['task_id'];
    projectId: EntityTaskIdentifierDTO['project_id'];
}

export type EntityIdentifier = EntityModelIdentifier | EntityTaskIdentifier;

export interface ConfigurableParametersComponents
    extends Omit<ConfigurableParametersComponentsDTO, 'groups' | 'parameters' | 'entity_identifier'> {
    entityIdentifier: EntityIdentifier;
    groups?: ConfigurableParametersGroups[];
    parameters?: ConfigurableParametersParams[];
}

export interface ConfigurableParametersTaskChain {
    taskId: ConfigurableParametersTaskChainDTO['task_id'];
    taskTitle: ConfigurableParametersTaskChainDTO['task_title'];
    components: ConfigurableParametersComponents[];
}